requirements.txt
playwright
pillow
imagehash
opencv-python
numpy
pytesseract
supabase
python-dotenv
requests
